+++ ecd-20200930/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-03-13 Version: 3.4.1
+- Update API ResetDesktops: add param DesktopGroupIds.
+- Update API ResetDesktops: add param ResetScope.
+
+
 2024-03-12 Version: 3.4.0
 - Support API DisconnectDesktopSessions.
 - Support API UnbindUserDesktop.